requirements.txt
PyQt6>=6.7.1
Pillow>=10.4.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
realesrgan>=0.3.0
basicsr>=1.4.2
noise>=1.2.2
PyOpenGL>=3.1.6
pygame-ce
google-generativeai
openai
